backend/requirements.txt
numpy>=1.21.0,<2.0.0
fasttext>=0.9.2
umap-learn>=0.5.3
requests>=2.25.0